Update volume model
