+++ api/types/model.go
@@ -139,29 +139,6 @@ type Exec struct {
 	Provider  Provider     `json:"provider"`
 }
 
-func NewExec(
-	ID string,
-	name string,
-	image string,
-	status Status,
-	createdAt time.Time,
-	region string,
-	provider Provider,
-	nodeMetadata *NodeMetadataV1,
-) *Exec {
-	return &Exec{
-		ID:        ID,
-		Name:      name,
-		Spec:      nodeMetadata.GetHardwareSpec(),
-		Image:     image,
-		Status:    status,
-		Network:   nodeMetadata.GetExecNetwork(),
-		CreatedAt: createdAt,
-		Region:    region,
-		Provider:  provider,
-	}
-}
-
 type ExecConfig struct {
 	Image   string         `json:"image"`
 	Command []string       `json:"command"`
@@ -197,4 +174,18 @@ func NewUserAccessToken(userID, tokenID, tokenName, displayText string, expiresA
 		DisplayText: displayText,
 		ExpiresAt:   expiresAt,
 	}
+}
+
+type Volume struct {
+	ID        string      `json:"id"`
+	Name      string      `json:"name"`
+	Size      int         `json:"size"`
+	State     VolumeState `json:"state"`
+	Provider  Provider    `json:"provider"`
+	ProjectID string      `json:"projectID"`
+}
+
+type VolumeState struct {
+	CreatedAt time.Time `json:"createdAt"`
+	UpdatedAt time.Time `json:"updatedAt"`
 }
